ENGR00262832 MX6SL-Lower Audio playback power.
Move MMDC to be sourced from PLL2_200M in audio mode.
Set the DDR freq to be 100MHz in audio mode.
Add code to drop DDR to 25MHz when ARM is in WFI while
playing audio. This will be the case when SDMA is transferring
data from the audio buffer in IRAM. Also float the DDR IO
pins in this state.

Set Audio PLL to bypass mode.
Source both WM8962 and SSI2 from audio PLL (PLL4).

Set AHB to 8MHz in Audio playback mode when ARM is going to enter WFI.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>