ov5645_mipi: remove bug in reset code and workaround intermittent I2C write problem
Removes a bug where the powerdown signal is sent low after reset.

Adds a workaround to retry i2c write failures once with a 1ms delay. This has
been seen to eliminate the intermittent problem with failures during the
firmware upload process during the device probe phase. We will be providing
an update to this after we properly identify and solve the issue.

Signed-off-by: John Weber <rjohnweber@gmail.com>