pico-imx7d: fix hanging on booting when enabling cortex-m4 on pico-imx7d with 512MB DRAM
The original memory layout is only suitable for pico-imx7d with 1GB DRAM.
The default offset address of RPMSG is 0xBFFF0000.
This offset address is out of 512MB memory.

In latest TechNexion i.mx u-boot with the commit:
--------------------------------------------
pico-imx7: reserve DDR memory for Cortex-M4
--------------------------------------------
We make a mechanism to detect DRAM size to configure the
memory layout for 512MB/1GB DRAM dynamically to solve this problem.